Move CommandLineUtils to their own module
