+++ CMakeLists.txt
@@ -203,6 +203,7 @@ if(WITH_GUI)
   add_subdirectory(src/CodeViewer)
   add_subdirectory(src/CodeViewerDemo)
   add_subdirectory(src/ConfigWidgets)
+  add_subdirectory(src/CommandLineUtils)
   add_subdirectory(src/DataViews)
   add_subdirectory(src/MoveFilesToDocuments)
   add_subdirectory(src/Orbit)
