+++ src/CommandLineUtils/include/CommandLineUtils/CommandLineUtils.h
@@ -2,12 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef _ORBIT_COMMAND_LINE_UTILS_H__
-#define _ORBIT_COMMAND_LINE_UTILS_H__
+#ifndef COMMAND_LINE_UTILS_COMMAND_LINE_UTILS_H__
+#define COMMAND_LINE_UTILS_COMMAND_LINE_UTILS_H__
 
 #include <QStringList>
 #include <string>
 #include <vector>
+
+namespace orbit_command_line_utils {
 
 // Extract command line flags by filtering the positional arguments out from the command line
 // arguments.
@@ -15,4 +17,6 @@ QStringList ExtractCommandLineFlags(const std::vector<std::string>& command_line
                                     const std::vector<char*>& positional_args);
 QStringList RemoveFlagsNotPassedToMainWindow(const QStringList& flags);
 
-#endif
+}  // namespace orbit_command_line_utils
+
+#endif  // COMMAND_LINE_UTILS_COMMAND_LINE_UTILS_H__
