+++ src/Orbit/CMakeLists.txt
@@ -8,17 +8,13 @@ add_executable(Orbit)
 
 target_sources(
     Orbit
-    PRIVATE CommandLineUtils.cpp
-            main.cpp
+    PRIVATE main.cpp
 )
 
-target_sources(
-  Orbit
-  PRIVATE CommandLineUtils.h
-)
 
 target_link_libraries(Orbit PRIVATE
         ClientFlags
+        CommandLineUtils
         MoveFilesToDocuments
         OrbitQt
         Style)
@@ -46,19 +42,3 @@ if(WIN32)
 endif()
 
 strip_symbols(Orbit)
-
-add_executable(OrbitMainTests)
-
-target_sources(OrbitMainTests PRIVATE
-               CommandLineUtils.cpp
-               CommandLineUtilsTest.cpp)
-
-target_link_libraries(
-  OrbitMainTests
-  PRIVATE ClientFlags
-          GTest::Main
-          MoveFilesToDocuments
-          OrbitQt
-          Style)
-
-register_test(OrbitMainTests)
