+++ src/Orbit/main.cpp
@@ -25,7 +25,7 @@
 #include <utility>
 #include <vector>
 
-#include "CommandLineUtils.h"
+#include "CommandLineUtils/CommandLineUtils.h"
 #include "MetricsUploader/ScopedMetric.h"
 #include "MoveFilesToDocuments/MoveFilesToDocuments.h"
 #include "OrbitBase/File.h"
@@ -203,7 +203,8 @@ int main(int argc, char* argv[]) {
   if (argc > 1) {
     command_line_args.assign(argv + 1, argv + argc);
   }
-  QStringList command_line_flags = ExtractCommandLineFlags(command_line_args, positional_args);
+  QStringList command_line_flags =
+      orbit_command_line_utils::ExtractCommandLineFlags(command_line_args, positional_args);
   // Skip program name in positional_args[0].
   std::vector<std::string> capture_file_paths(positional_args.begin() + 1, positional_args.end());
 
@@ -317,7 +318,8 @@ int main(int argc, char* argv[]) {
     QProcess::startDetached(orbit_executable, arguments);
   }
 
-  command_line_flags = RemoveFlagsNotPassedToMainWindow(command_line_flags);
+  command_line_flags =
+      orbit_command_line_utils::RemoveFlagsNotPassedToMainWindow(command_line_flags);
 
   RunUiInstance(deployment_configuration, &context.value(), command_line_flags, crash_handler.get(),
                 capture_file_paths.empty() ? "" : capture_file_paths[0]);
